Add interactive functionality.
This new functionality allows the user to type in the words one at a time,
getting instant results. It allows an endless number of words and is possibly
more convenient, but the words file is still helpful incase the search file
is misstyped. If you use interactive, you would have to retype all the words.
Theres not much waste of energy in using the normal mode, so I would recommend
using it over interactive.